feat: add eslint with airbnb config git hooks
