+++ .eslintrc
@@ -1,0 +1,6 @@
+{
+  "extends": "airbnb",
+  "env": {
+    "browser": true
+  }
+}
