+++ src/Product/ColorButton.jsx
@@ -1,5 +1,5 @@
-import React from 'react'
-import styled from 'styled-components'
+import React from 'react';
+import styled from 'styled-components';
 
 const ColorButton = styled.button`
   margin-right: 1rem;
@@ -9,12 +9,10 @@ const ColorButton = styled.button`
   border-radius: 50%;
   border: ${props => (props.active ? 'solid 1px #232122' : 'none')};
   background-color: ${props => props.value};
-`
+`;
 
-export default props => {
-  return (
-    <ColorButton value={props.value} active={props.active} type="button">
+export default props => (
+  <ColorButton value={props.value} active={props.active} type="button">
       Select {props.name} color
-    </ColorButton>
-  )
-}
+  </ColorButton>
+);
