+++ src/Product/DeliveryInfo.jsx
@@ -1,5 +1,5 @@
-import React from 'react'
-import styled from 'styled-components'
+import React from 'react';
+import styled from 'styled-components';
 
 const Container = styled.div`
   display: flex;
@@ -13,7 +13,7 @@ const Container = styled.div`
     align-items: center;
     margin-top: 2.5rem;
   }
-`
+`;
 
 const Title = styled.h2`
   display: none;
@@ -27,7 +27,7 @@ const Title = styled.h2`
   @media screen and (min-width: 48rem) {
     display: block;
   }
-`
+`;
 
 const Subtitle = styled.h3`
   margin: 0;
@@ -36,7 +36,7 @@ const Subtitle = styled.h3`
   font-size: .75rem;
   font-family: Raleway, 'Helvetica Neue', Helvetica, Arial;
   font-weight: bold;
-`
+`;
 
 const Text = styled.p`
   margin-bottom: 1.5rem !important;
@@ -46,7 +46,7 @@ const Text = styled.p`
   &:last-child {
     margin-bottom: 0 !important;
   }
-`
+`;
 
 const Image = styled.img`
   display: block;
@@ -58,36 +58,34 @@ const Image = styled.img`
   @media screen and (min-width: 48rem) {
     width: 100%;
   }
-`
+`;
 
-export default () => {
-  return (
-    <Container>
-      <div className="col-md-7">
-        <Image alt="Gift packaging demo" src="img/shipping-box.jpg" />
-      </div>
-      <div className="col-md-5 col-lg-4 col-lg-offset-1">
-        <Title>Delivery</Title>
-        <Subtitle>Free Next Day Delivery</Subtitle>
-        <Text>
+export default () => (
+  <Container>
+    <div className="col-md-7">
+      <Image alt="Gift packaging demo" src="img/shipping-box.jpg" />
+    </div>
+    <div className="col-md-5 col-lg-4 col-lg-offset-1">
+      <Title>Delivery</Title>
+      <Subtitle>Free Next Day Delivery</Subtitle>
+      <Text>
           Order before 7pm Monday to Thursday for delivery the next day
-        </Text>
+      </Text>
 
-        <Subtitle>Collect-in-Store</Subtitle>
-        <Text>
+      <Subtitle>Collect-in-Store</Subtitle>
+      <Text>
           Order online today and pick up your items in store as early as
           tomorrow
-        </Text>
+      </Text>
 
-        <Subtitle>Free Returns</Subtitle>
-        <Text>Enjoy free returns on your order</Text>
+      <Subtitle>Free Returns</Subtitle>
+      <Text>Enjoy free returns on your order</Text>
 
-        <Subtitle>Free Gift Packaging</Subtitle>
-        <Text>
+      <Subtitle>Free Gift Packaging</Subtitle>
+      <Text>
           Discover our gift packaging, a gold lined box tied with a coloured
           ribbon
-        </Text>
-      </div>
-    </Container>
-  )
-}
+      </Text>
+    </div>
+  </Container>
+);
