+++ src/Product/MainPanel.jsx
@@ -1,12 +1,12 @@
-import React from 'react'
-import styled from 'styled-components'
-import MediaQuery from 'react-responsive'
-import breakpoints from '../common/breakpoints'
-import PriceAndId from './PriceAndId'
-import ActionButtons from './ActionButtons'
-import ColorButton from './ColorButton'
-import Size from './Size'
-import SizeButton from './SizeButton'
+import React from 'react';
+import styled from 'styled-components';
+import MediaQuery from 'react-responsive';
+import breakpoints from '../common/breakpoints';
+import PriceAndId from './PriceAndId';
+import ActionButtons from './ActionButtons';
+import ColorButton from './ColorButton';
+import Size from './Size';
+import SizeButton from './SizeButton';
 
 const MainPanel = styled.section`
   margin-top: 1rem;
@@ -16,7 +16,7 @@ const MainPanel = styled.section`
     margin-top: 0;
     padding: 0;
   }
-`
+`;
 
 const Title = styled.h1`
   margin: 1rem .5rem;
@@ -35,7 +35,7 @@ const Title = styled.h1`
     margin: 0;
     margin-bottom: .5rem;
   }
-`
+`;
 
 const ColorName = styled.p`
   margin: 0;
@@ -50,7 +50,7 @@ const ColorName = styled.p`
   @media screen and (min-width: 62rem) {
     margin: 0;
   }
-`
+`;
 
 const ColorButtons = styled.div`
   margin-top: 1rem;
@@ -58,7 +58,7 @@ const ColorButtons = styled.div`
     margin: 0;
     margin-bottom: .5rem;
   }
-`
+`;
 
 const DeliveryTitle = styled.h3`
   margin-top: 1.5rem;
@@ -67,14 +67,14 @@ const DeliveryTitle = styled.h3`
   font-size: .75rem;
   font-family: Raleway, 'Helvetica Neue', Helvetica, Arial;
   font-weight: bold;
-`
+`;
 
 const DeliveryText = styled.p`
   margin: 0;
   line-height: 1.35;
   font-size: .75rem;
   font-family: Raleway, 'Helvetica Neue', Helvetica, Arial;
-`
+`;
 
 const Divider = styled.hr`
   margin-top: 2rem;
@@ -83,51 +83,49 @@ const Divider = styled.hr`
   @media screen and (min-width: 62rem) {
     display: none;
   }
-`
+`;
 
-export default () => {
-  return (
-    <MainPanel>
-      <MediaQuery minDeviceWidth={breakpoints.lg - 1}>
-        <Title>Long Cotton Gabardine Car Coat Coat Coat Coat Coat</Title>
-      </MediaQuery>
-      <PriceAndId />
-      <div className="row">
-        <div className="col-lg-6">
-          <ColorName>
+export default () => (
+  <MainPanel>
+    <MediaQuery minDeviceWidth={breakpoints.lg - 1}>
+      <Title>Long Cotton Gabardine Car Coat Coat Coat Coat Coat</Title>
+    </MediaQuery>
+    <PriceAndId />
+    <div className="row">
+      <div className="col-lg-6">
+        <ColorName>
             Colour: <b>Honey</b>
-          </ColorName>
-        </div>
-        <MediaQuery minDeviceWidth={breakpoints.lg - 1}>
-          <div className="col-lg-6">
-            <Size />
-          </div>
-        </MediaQuery>
+        </ColorName>
       </div>
-      <div className="row">
-        <div className="col-xs-12 col-lg-6">
-          <ColorButtons>
-            <ColorButton name="black" value="#232122" />
-            <ColorButton name="honey" value="#cfa880" active />
-          </ColorButtons>
-          <Divider />
-        </div>
-        <MediaQuery minDeviceWidth={breakpoints.lg - 1}>
-          <div className="col-lg-6">
-            <SizeButton>S</SizeButton>
-            <SizeButton>M</SizeButton>
-            <SizeButton>L</SizeButton>
-            <SizeButton>XL</SizeButton>
-          </div>
-        </MediaQuery>
-      </div>
-      <ActionButtons />
       <MediaQuery minDeviceWidth={breakpoints.lg - 1}>
-        <DeliveryTitle>Free Next Day Delivery</DeliveryTitle>
-        <DeliveryText>
-          Order before 7pm Monday to Thursday for delivery the next Day
-        </DeliveryText>
+        <div className="col-lg-6">
+          <Size />
+        </div>
       </MediaQuery>
-    </MainPanel>
-  )
-}
+    </div>
+    <div className="row">
+      <div className="col-xs-12 col-lg-6">
+        <ColorButtons>
+          <ColorButton name="black" value="#232122" />
+          <ColorButton name="honey" value="#cfa880" active />
+        </ColorButtons>
+        <Divider />
+      </div>
+      <MediaQuery minDeviceWidth={breakpoints.lg - 1}>
+        <div className="col-lg-6">
+          <SizeButton>S</SizeButton>
+          <SizeButton>M</SizeButton>
+          <SizeButton>L</SizeButton>
+          <SizeButton>XL</SizeButton>
+        </div>
+      </MediaQuery>
+    </div>
+    <ActionButtons />
+    <MediaQuery minDeviceWidth={breakpoints.lg - 1}>
+      <DeliveryTitle>Free Next Day Delivery</DeliveryTitle>
+      <DeliveryText>
+          Order before 7pm Monday to Thursday for delivery the next Day
+      </DeliveryText>
+    </MediaQuery>
+  </MainPanel>
+);
