+++ src/Product/RecommendationCard.jsx
@@ -1,17 +1,17 @@
-import React from 'react'
-import styled from 'styled-components'
-import { FormattedNumber } from 'react-intl'
+import React from 'react';
+import styled from 'styled-components';
+import { FormattedNumber } from 'react-intl';
 
 const RecommendationCard = styled.a`
   display: block;
   margin-bottom: 2rem;
-`
+`;
 
 const Image = styled.img`
   display: block;
   width: 100%;
   height: auto;
-`
+`;
 
 const Name = styled.h3`
   margin-top: .5rem;
@@ -25,7 +25,7 @@ const Name = styled.h3`
     font-family: Raleway, 'Helvetica Neue', Helvetica, Arial;
     font-weight: 600;
   }
-`
+`;
 
 const Price = styled.h5`
   margin: 0;
@@ -37,24 +37,22 @@ const Price = styled.h5`
   @media screen and (min-width: 62rem) {
     line-height: 1.4;
   }
-`
+`;
 
-export default props => {
-  return (
-    <RecommendationCard>
-      <Image alt="" src={props.image} />
-      <Name>
-        {props.name}
-      </Name>
-      <Price>
-        <FormattedNumber
-          value={props.price}
-          style="currency"
-          currency={props.currency}
-          currencyDisplay="symbol"
-          minimumFractionDigits={0}
-        />
-      </Price>
-    </RecommendationCard>
-  )
-}
+export default props => (
+  <RecommendationCard>
+    <Image alt="" src={props.image} />
+    <Name>
+      {props.name}
+    </Name>
+    <Price>
+      <FormattedNumber
+        value={props.price}
+        style="currency"
+        currency={props.currency}
+        currencyDisplay="symbol"
+        minimumFractionDigits={0}
+      />
+    </Price>
+  </RecommendationCard>
+);
