+++ src/Product/Size.jsx
@@ -1,11 +1,11 @@
-import React from 'react'
-import styled from 'styled-components'
+import React from 'react';
+import styled from 'styled-components';
 
 const Container = styled.div`
   display: flex;
   justify-content: space-between;
   margin-bottom: 1rem;
-`
+`;
 
 const Size = styled.p`
   margin: 0;
@@ -13,7 +13,7 @@ const Size = styled.p`
   font-family: 'Raleway';
   color: #171717;
   line-height: 1.33;
-`
+`;
 
 const TextButton = styled.button`
   padding: 0;
@@ -23,15 +23,13 @@ const TextButton = styled.button`
   border: 0;
   background-color: inherit;
   color: #171717;
-`
+`;
 
-export default () => {
-  return (
-    <Container>
-      <Size>
+export default () => (
+  <Container>
+    <Size>
         Size: <b>XL</b>
-      </Size>
-      <TextButton>Need size help?</TextButton>
-    </Container>
-  )
-}
+    </Size>
+    <TextButton>Need size help?</TextButton>
+  </Container>
+);
