+++ src/common/breakpoints.jsx
@@ -2,5 +2,5 @@ export default {
   sm: 576,
   md: 768,
   lg: 992,
-  xl: 1200
-}
+  xl: 1200,
+};
