+++ src/index.js
@@ -1,10 +1,11 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import 'normalize.css'
-import 'flexboxgrid2'
-import './index.css'
-import App from './App'
-import registerServiceWorker from './registerServiceWorker'
+/* eslint-disable react/jsx-filename-extension */
+import React from 'react';
+import ReactDOM from 'react-dom';
+import 'normalize.css';
+import 'flexboxgrid2/dist/flexboxgrid.css';
+import './index.css';
+import App from './App';
+import registerServiceWorker from './registerServiceWorker';
 
-ReactDOM.render(<App />, document.getElementById('root'))
-registerServiceWorker()
+ReactDOM.render(<App />, document.getElementById('root'));
+registerServiceWorker();
